ai-services/id-verification/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.22.0
redis==4.5.5
python-dotenv==1.0.0
pillow==10.0.0
deepface==0.0.95
python-multipart==0.0.6
opencv-python==4.9.0.80
numpy==1.26.4
tf-keras
